backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for employee management system."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "psycopg (>=3.2.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
